backend/pyproject.toml
[project]
name = "smart-trainer-backend"
version = "0.1.0"
description = "AI Trainer backend - FastAPI, Intervals.icu, Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-generativeai>=0.4.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "cryptography>=42.0.0",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.uv]
dev-dependencies = []
